requirements.txt
art==5.3
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.46.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
exif==1.3.3
exif-delete==0.9.8
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient @ file:///C:/Users/Samir/Documents/tvf/mysqlclient-1.4.6-cp37-cp37m-win32.whl
oauthlib==3.1.1
packaging==21.3
pika==1.2.0
Pillow==9.0.1
plum-py==0.7.3
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
schedule==1.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
